Windows fixes. Stopped weird issue with loading images by not doing it all at once.

+++ engine/src/io/file.hpp
@@ -4,7 +4,7 @@
 #include <utility>
 
 namespace smeg {
-	typedef std::pair<std::unique_ptr<unsigned char[]>, int> FileData;
+	typedef std::pair<std::unique_ptr<unsigned char[]>, long> FileData;
 	class File {
 		public:
 			static FileData Load(const char* filePath);
